add a separate folder for the different message types
