+++ lib/s3i_flutter.dart
@@ -11,8 +11,8 @@ export 'src/broker/attribute_value_messages.dart';
 export 'src/broker/broker_interfaces.dart';
 export 'src/broker/broker_rest_connector.dart'
     if (dart.library.io) 'src/broker/broker_amqp_connector.dart';
-export 'src/broker/service_messages.dart';
-export 'src/broker/user_message.dart';
+export 'src/broker/messages/service_messages.dart';
+export 'src/broker/messages/user_message.dart';
 
 //directory
 export 'src/directory/dir_object.dart';
