+++ lib/src/broker/broker_rest_connector.dart
@@ -4,7 +4,7 @@ import 'package:http/http.dart';
 import 'package:s3i_flutter/src/auth/authentication_manager.dart';
 import 'package:s3i_flutter/src/auth/tokens.dart';
 import 'package:s3i_flutter/src/broker/broker_interfaces.dart';
-import 'package:s3i_flutter/src/broker/message.dart';
+import 'package:s3i_flutter/src/broker/messages/message.dart';
 import 'package:s3i_flutter/src/exceptions/network_response_exception.dart';
 import 'package:s3i_flutter/src/exceptions/s3i_exception.dart';
