+++ lib/src/broker/messages/message.dart
